change to arg input
